Timer and Counter Fixed
Timer adjusted to make the solenoid work effectively and return to initial position. Counter counts the number of the plastic rings in the chute, exceeding 5 will disable the sorting solenoid and will allow the plastic ring to pass through the sensing (belt) conveyor and will be rejected by the 'reject' solenoid. Since the component is not assembled i.e., no metal peg. Solenoid error safety precaution is also implemented